+++ tests/x86-64-level.sh
@@ -255,6 +255,39 @@ for level in 1.2 world; do
 done
 
 
+# x86-64-level <unknown option>
+for option in --foo=1 --ass--ert=2 --assert--=2 --bar; do
+    echo "* x86-64-level ${option} (exception)"
+    stderr=$(x86-64-level "${option}" 2>&1)
+    exit_code=$?
+    if [[ ${exit_code} -eq 0 ]]; then
+        >&2 echo "ERROR: Exit code should be non-zero: ${exit_code}"
+        nerrors=$((nerrors + 1))
+    fi
+
+    if [[ -z ${stderr} ]]; then
+        >&2 echo "ERROR: No error message: '${stderr}'"
+        nerrors=$((nerrors + 1))
+    fi
+
+    if ! head -n 1 <<< "${stderr}" | grep -q -E "^ERROR:"; then
+        >&2 echo "ERROR: Standard error output does not begin with 'ERROR:': '${stderr}'"
+        nerrors=$((nerrors + 1))
+    fi
+
+    if ! grep -q -F "ERROR: Unknown option: ${option}" <<< "${stderr}"; then
+        >&2 echo "ERROR: Unexpected error message: '${stderr}'"
+        nerrors=$((nerrors + 1))
+    fi
+
+    ## Outputs nothing to stdout
+    stdout=$(x86-64-level "${option}" 2> /dev/null)
+    if [[ -n ${stdout} ]]; then
+        >&2 echo "ERROR: Detected output to standard output: ${stdout}"
+        nerrors=$((nerrors + 1))
+    fi
+done
+
 
 # x86-64-level --assert=''
 echo "* x86-64-level --assert='' (exception)"